break out ex3 into several files
