--- temporal102/ex3/cabal.project
@@ -1,1 +1,0 @@
-packages: .
